# Update miniAOD AK8 userFloat content. Add PUPPI substructure. Remove CMSTT.