--- a/RecoJets/JetProducers/python/ak8PFJetsPuppi_cfi.py
+++ b/RecoJets/JetProducers/python/ak8PFJetsPuppi_cfi.py
@@ -1,0 +1,10 @@
+import FWCore.ParameterSet.Config as cms
+
+from RecoJets.JetProducers.PFJetParameters_cfi import *
+from RecoJets.JetProducers.AnomalousCellParameters_cfi import *
+from RecoJets.JetProducers.ak4PFJetsPuppi_cfi import ak4PFJetsPuppi
+
+ak8PFJetsPuppi = ak4PFJetsPuppi.clone(
+    rParam = 0.8    
+    )
+