We have the following proposal for the AK8chs jet MiniAOD userfloat content in 80X/81X:
- Remove unused run 1 taggers (CMS Top tagger)
- Add PUPPI based taggers (PUPPI soft drop + PUPPI Nsubjettiness ) for matched AK8 PUPPI jets
- In order to save space, add only PUPPI subjets and userfloats for matched PUPPI jet 4-vector and substructure variables.

Size/timing estimates are given in the slides linked here.
https://indico.cern.ch/event/510570/contribution/7/attachments/1246782/1836484/160321_-_Jet_substructure_MiniAOD_content.pptx.pdf

Cheers,
@rappoccio @ahinzmann @jdolen