@cmsbuild 2016-03-21
A new Pull Request was created by  @jdolen (James Dolen) for CMSSW_8_1_X.

It involves the following packages:

CommonTools/RecoAlgos
PhysicsTools/PatAlgos
RecoJets/Configuration
RecoJets/JetProducers


@cvuosalo, @monttj, @cmsbuild, @slava77, @vadler, @davidlange6 can you please review it and eventually sign? Thanks.
@rappoccio, @imarches, @makortel, @acaudron, @abbiendi, @mmarionncern, @jhgoh, @ahinzmann, @nhanvtran, @yslai, @schoef, @ferencek, @gpetruc, @mariadalfonso, @pvmulder, @TaiSakuma this is something you requested to watch as well.
@slava77, @degano, @smuzaffar you are the release manager for this.

cms-bot commands are list here cms-sw/cmssw#13028