@slava77 2016-03-21
@cmsbuild please test