@cmsbuild 2016-03-21
The tests are being triggered in jenkins.