@cmsbuild 2016-03-21
+1
Tested at: 85742010a39d5dfb8000f85ffb190c0e926c7f34
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-13793/11995/summary.html