@cvuosalo 2016-03-23
An extended test of workflow 25202.0_TTbar_13 with 70 events  and no DQM/Validation step against baseline CMSSW_8_1_X_2016-03-14-1100 shows the following time and memory usage. TopTagCHS modules are removed and replaced by AK8PFPuppi modules. The removed modules had a total time of about 50 ms/evt, while the new modules together only take about 40 ms/evt. Overall, there is no significant change in CPU time for RECO.

```
Baseline
Max VSIZ 2280.05 on evt 36 ; max RSS 1826.91 on evt 36
Time av 15.1361 s/evt   max 48.7359 s on evt 1
M1 Time av 14.6492 s/evt   max 27.0861 s on evt 36
M8 Time av 14.8417 s/evt   max 27.0861 s on evt 36

PR 13793
Summary for 70 events
Max VSIZ 2276.84 on evt 36 ; max RSS 1765.72 on evt 61
Time av 15.123 s/evt   max 50.7816 s on evt 1
M1 Time av 14.6063 s/evt   max 27.0756 s on evt 36
M8 Time av 14.7225 s/evt   max 27.0756 s on evt 36

  delta/mean delta/orJob     original                   new       module name
  ---------- ------------     --------                  ----       ------------
     removed      -0.00%         0.07 ms/ev ->         0.00 ms/ev patJetPartonMatchCMSTopTagCHS
     removed      -0.01%         1.77 ms/ev ->         0.00 ms/ev pfSecondaryVertexTagInfosCMSTopTagCHSSubjets
     removed      -0.00%         0.08 ms/ev ->         0.00 ms/ev patJetCorrFactorsCMSTopTagCHS
     removed      -0.00%         0.39 ms/ev ->         0.00 ms/ev patJetsCMSTopTagCHS
     removed      -0.00%         0.04 ms/ev ->         0.00 ms/ev caTopTagInfos
     removed      -0.00%         0.03 ms/ev ->         0.00 ms/ev patJetGenJetMatchCMSTopTagCHS
     removed      -0.01%         1.24 ms/ev ->         0.00 ms/ev patJetFlavourAssociationCMSTopTagCHSSubjets
     removed      -0.01%         1.02 ms/ev ->         0.00 ms/ev pfInclusiveSecondaryVertexFinderTagInfosCMSTopTagCHSSubjets
     removed      -0.00%         0.03 ms/ev ->         0.00 ms/ev patJetGenJetMatchCMSTopTagCHSSubjets
     removed      -0.00%         0.07 ms/ev ->         0.00 ms/ev ak8PFJetsCHSTrimmedMass
     removed      -0.01%         1.43 ms/ev ->         0.00 ms/ev patJetsCMSTopTagCHSSubjets
     removed      -0.00%         0.04 ms/ev ->         0.00 ms/ev ak8PFJetsCHSFilteredMass
     removed      -0.00%         0.06 ms/ev ->         0.00 ms/ev selectedPatJetsCMSTopTagCHS
     removed      -0.00%         0.04 ms/ev ->         0.00 ms/ev caTopTagInfosPAT
     removed      -0.00%         0.11 ms/ev ->         0.00 ms/ev selectedPatJetsCMSTopTagCHSSubjets
     removed      -0.00%         0.23 ms/ev ->         0.00 ms/ev pfCombinedSecondaryVertexV2BJetTagsCMSTopTagCHSSubjets
     removed      -0.00%         0.08 ms/ev ->         0.00 ms/ev slimmedJetsCMSTopTagCHSSubjets
     removed      -0.00%         0.13 ms/ev ->         0.00 ms/ev patJetPartonMatchCMSTopTagCHSSubjets
     removed      -0.01%         1.72 ms/ev ->         0.00 ms/ev pfImpactParameterTagInfosCMSTopTagCHSSubjets
     removed      -0.00%         0.04 ms/ev ->         0.00 ms/ev slimmedJetsCMSTopTagCHSPacked
     removed      -0.04%         4.98 ms/ev ->         0.00 ms/ev ak8PFJetsCHSFiltered
     removed      -0.00%         0.13 ms/ev ->         0.00 ms/ev patJetCorrFactorsCMSTopTagCHSSubjets
     removed      -0.29%        41.15 ms/ev ->         0.00 ms/ev ak8PFJetsCHSTrimmed
     removed      -0.00%         0.23 ms/ev ->         0.00 ms/ev pfCombinedInclusiveSecondaryVertexV2BJetTagsCMSTopTagCHSSubjets
      added      +0.00%         0.00 ms/ev ->         0.03 ms/ev patJetGenJetMatchAK8PFPuppiSoftDropSubjets
      added      +0.00%         0.00 ms/ev ->         0.06 ms/ev patJetPartonMatchAK8PFPuppiSoftDrop
      added      +0.00%         0.00 ms/ev ->         0.16 ms/ev pfCombinedInclusiveSecondaryVertexV2BJetTagsAK8PFPuppiSoftDropSubjets
      added      +0.01%         0.00 ms/ev ->         1.31 ms/ev patJetsAK8PFPuppiSoftDropSubjets
      added      +0.00%         0.00 ms/ev ->         0.32 ms/ev patJetPartonMatchAK8Puppi
      added      +0.00%         0.00 ms/ev ->         0.38 ms/ev patJetsAK8PFPuppiSoftDrop
      added      +0.00%         0.00 ms/ev ->         0.07 ms/ev selectedPatJetsAK8PFPuppiSoftDrop
      added      +0.00%         0.00 ms/ev ->         0.16 ms/ev pfCombinedSecondaryVertexV2BJetTagsAK8PFPuppiSoftDropSubjets
      added      +0.00%         0.00 ms/ev ->         0.05 ms/ev ak8PFJetsPuppiSoftDropMass
      added      +0.00%         0.00 ms/ev ->         0.08 ms/ev patJetCorrFactorsAK8PFPuppiSoftDropSubjets
      added      +0.00%         0.00 ms/ev ->         0.08 ms/ev patJetPartonMatchAK8PFPuppiSoftDropSubjets
      added      +0.00%         0.00 ms/ev ->         0.07 ms/ev patJetCorrFactorsAK8PFPuppiSoftDrop
      added      +0.19%         0.00 ms/ev ->        27.16 ms/ev ak8PFJetsPuppi
      added      +0.00%         0.00 ms/ev ->         0.09 ms/ev selectedPatJetsAK8PFPuppiSoftDropSubjets
      added      +0.00%         0.00 ms/ev ->         0.59 ms/ev patJetFlavourAssociationAK8PFPuppiSoftDropSubjets
      added      +0.00%         0.00 ms/ev ->         0.04 ms/ev slimmedJetsAK8PFPuppiSoftDropPacked
      added      +0.00%         0.00 ms/ev ->         0.42 ms/ev pfInclusiveSecondaryVertexFinderTagInfosAK8PFPuppiSoftDropSubjets
      added      +0.06%         0.00 ms/ev ->         7.90 ms/ev NjettinessAK8Puppi
      added      +0.00%         0.00 ms/ev ->         0.03 ms/ev patJetGenJetMatchAK8PFPuppiSoftDrop
      added      +0.00%         0.00 ms/ev ->         0.04 ms/ev patJetGenJetMatchAK8Puppi
      added      +0.00%         0.00 ms/ev ->         0.05 ms/ev slimmedJetsAK8PFPuppiSoftDropSubjets
      added      +0.00%         0.00 ms/ev ->         0.32 ms/ev patJetCorrFactorsAK8Puppi
      added      +0.00%         0.00 ms/ev ->         0.19 ms/ev ak8PFJetsPuppiValueMap
      added      +0.01%         0.00 ms/ev ->         1.07 ms/ev pfImpactParameterTagInfosAK8PFPuppiSoftDropSubjets
      added      +0.01%         0.00 ms/ev ->         1.10 ms/ev pfSecondaryVertexTagInfosAK8PFPuppiSoftDropSubjets
      added      +0.01%         0.00 ms/ev ->         1.20 ms/ev patJetsAK8Puppi
      added      +0.00%         0.00 ms/ev ->         0.67 ms/ev ak8PFJetsPuppiSoftDrop
      added      +0.00%         0.00 ms/ev ->         0.56 ms/ev patJetFlavourAssociationAK8Puppi
```