@cvuosalo 2016-03-23
@jdolen : Reco is ready to approve this PR when the little code fixes requested by @slava77 and @ferencek are completed.