@cvuosalo 2016-03-24
@cmsbuild please test