@cvuosalo 2016-03-25
Additional, possibly more accurate, test results running only the PAT step with 1000 events to produce Mini-AOD were obtained with this command:
```
cmsDriver.py miniAOD-prod -s PAT --eventcontent MINIAODSIM --runUnscheduled --mc --filein /store/relval/CMSSW_8_1_0_pre1/RelValTTbar_13/GEN-SIM-RECO/PU25ns_80X_mcRun2_asymptotic_v6-v1/10000/407BD44A-DCE8-E511-BF44-0025905A60B6.root --conditions 80X_mcRun2_asymptotic_v6 -n 1000 --customise Validation/Performance/TimeMemoryInfo.py
```
Wall-clock time and memory usage are shown, along with a second timing measure that excludes the first event. Only the largest time changes are listed (see above for full listing of added and removed modules).
```
Baseline
Max VSIZ 1736.16 on evt 910 ; max RSS 1267.02 on evt 993
Time av 5.51691 s/evt   max 112.822 s on evt 778
M1 Time av 5.48936 s/evt   max 112.822 s on evt 778
M8 Time av 5.51286 s/evt   max 112.822 s on evt 778

PR 13793
Max VSIZ 1704.7 on evt 401 ; max RSS 1273.86 on evt 666
Time av 5.47937 s/evt   max 115.781 s on evt 778
M1 Time av 5.4533 s/evt   max 115.781 s on evt 778
M8 Time av 5.47663 s/evt   max 115.781 s on evt 778

  delta/mean delta/orJob     original                   new       module name
  ---------- ------------     --------                  ----       ------------
     removed      -0.07%         3.35 ms/ev ->         0.00 ms/ev pfSecondaryVertexTagInfosCMSTopTagCHSSubjets
     removed      -0.07%         3.39 ms/ev ->         0.00 ms/ev patJetFlavourAssociationCMSTopTagCHSSubjets
     removed      -0.07%         3.48 ms/ev ->         0.00 ms/ev pfInclusiveSecondaryVertexFinderTagInfosCMSTopTagCHSSubjets
     removed      -0.17%         8.75 ms/ev ->         0.00 ms/ev patJetsCMSTopTagCHSSubjets
     removed      -0.07%         3.70 ms/ev ->         0.00 ms/ev pfImpactParameterTagInfosCMSTopTagCHSSubjets
     removed      -0.32%        16.43 ms/ev ->         0.00 ms/ev ak8PFJetsCHSFiltered
     removed      -2.00%       101.57 ms/ev ->         0.00 ms/ev ak8PFJetsCHSTrimmed
       added      +0.08%         0.00 ms/ev ->         3.89 ms/ev patJetsAK8PFPuppiSoftDropSubjets
       added      +1.02%         0.00 ms/ev ->        51.86 ms/ev ak8PFJetsPuppi
       added      +0.44%         0.00 ms/ev ->        22.64 ms/ev NjettinessAK8Puppi
       added      +0.06%         0.00 ms/ev ->         3.14 ms/ev pfSecondaryVertexTagInfosAK8PFPuppiSoftDropSubjets
       added      +0.08%         0.00 ms/ev ->         4.29 ms/ev patJetsAK8Puppi
   +0.173479      +0.42%       111.30 ms/ev ->       132.44 ms/ev pfCandidateToVertexAssociationPuppi
   -0.134996      -0.80%       320.81 ms/ev ->       280.24 ms/ev puppi
  ---------- ------------     --------                  ----       ------------
Job total:  5.09096 s/ev ==> 5.03398 s/ev

```
Change in the size of Mini-AOD:
```-----------------------------------------------------------------
   or, B         new, B      delta, B   delta, %   deltaJ, %    branch 
-----------------------------------------------------------------
    307.6 ->       317.7         10      3.2   0.02     patJets_slimmedJetsAK8__PAT.
     60.6 ->         0.0        -61     OLDO  -0.13     recoCATopJetTagInfos_caTopTagInfosPAT__PAT.
    293.3 ->         0.0       -293     OLDO  -0.63     patJets_slimmedJetsCMSTopTagCHSPacked_SubJets_PAT.
      0.0 ->       302.1        302     NEWO   0.65     patJets_slimmedJetsAK8PFPuppiSoftDropPacked_SubJets_PAT.
-------------------------------------------------------------
    46279 ->       46237        -42            -0.1     ALL BRANCHES
```